LPS-170680 Extract the settings for User Notification to its own component
